[Refactor] useBrowserTabTitle composable
